Deleted the mispaced scripts. Made further changes with the LFS parser

+++ lfsparser/src/main/java/lfsparser/Parser.java
@@ -19,6 +19,30 @@ public class Parser {
 		document = Jsoup.parse(Util.readFileSpecial(documentPath));
 	}
 
+	public Document getDocument() {
+		return document;
+	}
+
+	public void setDocument(Document document) {
+		this.document = document;
+	}
+
+	public String getTarball() {
+		return tarball;
+	}
+
+	public void setTarball(String tarball) {
+		this.tarball = tarball;
+	}
+
+	public void setCommands(String commands) {
+		this.commands = commands;
+	}
+
+	public void setStepName(String stepName) {
+		this.stepName = stepName;
+	}
+
 	public void getTarball(List<String> tarballs) {
 		String title = document.select("title").first().html().replace("&nbsp;", " ").replace("br3ak", "").trim()
 				.toLowerCase();
@@ -48,6 +72,13 @@ public class Parser {
 		if (index != -1) {
 			this.tarball = tarballs.get(index);
 		}
+		if (this.tarball == null && this.stepName.contains("XML::Parser")) {
+			for (String str : tarballs) {
+				if (str.contains("XML-Parser")) {
+					this.tarball = str;
+				}
+			}
+		}
 	}
 
 	public void getCommands() {
@@ -59,8 +90,9 @@ public class Parser {
 		this.commands = builder.toString();
 	}
 
-	public void getStepName() {
+	public String getStepName() {
 		this.stepName = document.select("title").html().replace("&nbsp;", " ").replace("br3ak", "").trim();
+		return this.stepName;
 	}
 
 	public void parse(List<String> tarballs) {
@@ -74,18 +106,24 @@ public class Parser {
 		builder.append("#!/bin/bash\n");
 		builder.append("set -e\n");
 		builder.append("set +h\n\n");
-		builder.append(". build.conf\n\n");
-		builder.append("TARBALL=" + tarball + "\n\n");
+		builder.append(". /sources/build.conf\n\n");
+		if (tarball != null) {
+			builder.append("TARBALL=" + tarball + "\n\n");
+		}
 		builder.append("STEPNAME=\"" + stepName + "\"\n\n");
 		builder.append("if ! grep \"$STEPNAME\" $BUILD_LOG\n");
 		builder.append("then\n\n");
 		builder.append("cd $SOURCE_DIR\n");
-		builder.append("DIRECTORY=`tar tf $TARBALL | cut -d/ -f1 | uniq`\n\n");
-		builder.append("tar xf $TARBALL\n");
-		builder.append("cd $DIRECTORY\n");
+		if (tarball != null) {
+			builder.append("DIRECTORY=`tar tf $TARBALL | cut -d/ -f1 | uniq`\n\n");
+			builder.append("tar xf $TARBALL\n");
+			builder.append("cd $DIRECTORY\n");
+		}
 		builder.append("\n" + commands + "\n");
 		builder.append("cd $SOURCE_DIR\n");
-		builder.append("rm -rf $DIRECTORY\n");
+		if (tarball != null) {
+			builder.append("rm -rf $DIRECTORY\n");
+		}
 		builder.append("rm -rf gcc-build\n");
 		builder.append("rm -rf glibc-build\n");
 		builder.append("rm -rf binutils-build\n");
